507097 - Fixed guest provisioning virt settings
Fixed guest provisioning  to correctly set the virt params in cobbler
system record.

+++ java/code/src/org/cobbler/SystemRecord.java
@@ -284,14 +284,14 @@ public class SystemRecord extends CobblerObject {
       /**
       * @param virtFileSizeIn the VirtFileSize
       */
-      public void  setVirtFileSize(long virtFileSizeIn) {
+      public void  setVirtFileSize(int virtFileSizeIn) {
           modify(VIRT_FILE_SIZE, virtFileSizeIn);
       }
 
       /**
       * @param virtRamIn the VirtRam
       */
-      public void  setVirtRam(long virtRamIn) {
+      public void  setVirtRam(int virtRamIn) {
           modify(VIRT_RAM, virtRamIn);
       }
